Refactoring kernel sources.
